chore: move api-decorator tests to its own file
